requirements.txt
fastapi 
uvicorn 
python-multipart 
anthropic 
pypdf2 
python-dotenv 
pytest 
pytest-asyncio 
httpx
pycryptodome